requirements.txt
numpy==1.26.1
opencv-contrib-python==4.8.1.78
Pillow==10.1.0
psycopg2==2.9.9
python-dotenv==1.0.0
